ref: moved routes to the route group, changed sidebar background and button background color
