--- app/(root)/dashboard/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const page = () => {
-  return (
-    <div>
-      hello home
-    </div>
-  )
-}
-
-export default page
