+++ app/(routes)/(dashboard)/sales/page.tsx
