+++ app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from "next";
 import "./globals.css";
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Nexus",
+  description: "Real Estate CRM Software",
 };
 
 export default function RootLayout({
