+++ components/composites/Sidebar.tsx
@@ -2,16 +2,18 @@
 
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { sidebarLinks } from "@/data";
+import { sidebarLinks } from "@/app/(lead-feat)/data";
 import Text from "../ui/Text";
+
+
 
 const Sidebar = () => {
   const pathname = usePathname();
 
-  const isPathActive = (path: string) => pathname === path;
+  const isPathActive = (path: string) => pathname === path
 
   return (
-    <div className="w-64 bg-black/5 min-h-screen text-black">
+    <div className="w-64 bg-dark-bg min-h-screen">
       {sidebarLinks.map((link, index) => {
         const isActive = link.path && isPathActive(link.path);
 
@@ -22,11 +24,14 @@ const Sidebar = () => {
                 <div
                   className={`${
                     isActive
-                      ? "bg-blue-400/20 text-blue-700 border-r-4 font-semibold border-blue-700"
-                      : "text-black/80"
-                  } flex items-center p-2 cursor-pointer`}
+                      ? " bg-blue-700 text-white border-r-4 border-blue-700"
+                      : "text-white"
+                  } flex items-center p-2 cursor-pointer text-sm`}
                 >
-                  <Text variant="span" className="ml-4">{link.name}</Text>
+                  {link.icon}
+                  <Text variant="span" className="ml-4">
+                    {link.name}
+                  </Text>
                 </div>
               </Link>
             )}
