+++ tailwind.config.ts
@@ -10,6 +10,8 @@ const config: Config = {
   theme: {
   	extend: {
   		colors: {
+			"dark-bg": "#1B1D2A",
+			"blue-color": "#325AE7",
   			background: 'hsl(var(--background))',
   			foreground: 'hsl(var(--foreground))',
   			card: {
